+++ ide-plugins/jetbrains/src/main/java-intellij/com/mercedesbenz/sechub/plugin/idea/sechubaccess/SecHubAccessFactory.java
@@ -1,8 +1,6 @@
 // SPDX-License-Identifier: MIT
 package com.mercedesbenz.sechub.plugin.idea.sechubaccess;
 
-import java.util.Objects;
-
 import com.intellij.credentialStore.Credentials;
 import com.mercedesbenz.sechub.settings.SechubSettings;
 import com.mercedesbenz.sechub.settings.SechubSettingsCredentialsSupport;
@@ -11,8 +9,23 @@ public class SecHubAccessFactory {
     public static SecHubAccess create() {
         SechubSettingsCredentialsSupport sechubSettingsCredentialsSupport = new SechubSettingsCredentialsSupport();
         Credentials credentials = sechubSettingsCredentialsSupport.retrieveCredentials();
-        SechubSettings.State state = Objects.requireNonNull(SechubSettings.getInstance().getState());
-
-        return new SecHubAccess(state.serverURL, credentials.getUserName(), credentials.getPasswordAsString(), true);
+        SechubSettings.State state = SechubSettings.getInstance().getState();
+        if (isInvalidState(state, credentials)) {
+            return new NoOpSecHubAccessClient();
+        }
+        return new SecHubAccessClient(state.serverURL, credentials.getUserName(), credentials.getPasswordAsString(), state.sslTrustAll);
     }
+
+    private static boolean isInvalidState(SechubSettings.State state, Credentials credentials) {
+        /* @formatter:off */
+        return state == null ||
+                state.serverURL == null ||
+                state.serverURL.isBlank() ||
+                credentials == null ||
+                credentials.getUserName() == null ||
+                credentials.getPassword() == null;
+        /* @formatter:on */
+    }
+
+
 }